+++ test/static_unordered_flat_set.cpp
@@ -493,6 +493,33 @@ void test_static_unordered_flat_set()
         }
     }
 
+    PRINT("Test insert(K&&)");
+    {
+        sfl::static_unordered_flat_set<xobj, 100, xobj::equal> set;
+
+        {
+            CHECK(set.insert(10) == std::make_pair(set.nth(0), true));
+            CHECK(set.insert(20) == std::make_pair(set.nth(1), true));
+            CHECK(set.insert(30) == std::make_pair(set.nth(2), true));
+
+            CHECK(set.size() == 3);
+            CHECK(set.nth(0)->value() == 10);
+            CHECK(set.nth(1)->value() == 20);
+            CHECK(set.nth(2)->value() == 30);
+        }
+
+        {
+            CHECK(set.insert(10) == std::make_pair(set.nth(0), false));
+            CHECK(set.insert(20) == std::make_pair(set.nth(1), false));
+            CHECK(set.insert(30) == std::make_pair(set.nth(2), false));
+
+            CHECK(set.size() == 3);
+            CHECK(set.nth(0)->value() == 10);
+            CHECK(set.nth(1)->value() == 20);
+            CHECK(set.nth(2)->value() == 30);
+        }
+    }
+
     PRINT("Test insert(const_iterator, const value_type&)");
     {
         sfl::static_unordered_flat_set<xint_xint, 100, std::equal_to<xint_xint>> set;
@@ -580,6 +607,33 @@ void test_static_unordered_flat_set()
             CHECK(value_10_2.first == +10); CHECK(value_10_2.second == +2);
             CHECK(value_20_2.first == +20); CHECK(value_20_2.second == +2);
             CHECK(value_30_2.first == +30); CHECK(value_30_2.second == +2);
+        }
+    }
+
+    PRINT("Test insert(const_iterator, K&&)");
+    {
+        sfl::static_unordered_flat_set<xobj, 100, xobj::equal> set;
+
+        {
+            CHECK(set.insert(set.begin(), 10) == set.nth(0));
+            CHECK(set.insert(set.begin(), 20) == set.nth(1));
+            CHECK(set.insert(set.begin(), 30) == set.nth(2));
+
+            CHECK(set.size() == 3);
+            CHECK(set.nth(0)->value() == 10);
+            CHECK(set.nth(1)->value() == 20);
+            CHECK(set.nth(2)->value() == 30);
+        }
+
+        {
+            CHECK(set.insert(set.begin(), 10) == set.nth(0));
+            CHECK(set.insert(set.begin(), 20) == set.nth(1));
+            CHECK(set.insert(set.begin(), 30) == set.nth(2));
+
+            CHECK(set.size() == 3);
+            CHECK(set.nth(0)->value() == 10);
+            CHECK(set.nth(1)->value() == 20);
+            CHECK(set.nth(2)->value() == 30);
         }
     }